backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
pandas
openpyxl
numpy
celery
redis
python-decouple
gunicorn
ipython
Pillow